move MustParseBool func
