+++ internal/types/types_test.go
@@ -1,4 +1,4 @@
-package config
+package types
 
 import (
 	"testing"
